+++ deploy/sbom/resolve_licenses_test.py
@@ -8,7 +8,39 @@ import threading
 import time
 from concurrent.futures import ThreadPoolExecutor
 
-from resolve_licenses import _last_request, _rate_limit, _rate_lock
+from resolve_licenses import _last_request, _rate_limit, _rate_lock, needs_fix
+
+
+def test_empty_licenses_needs_fix() -> None:
+    assert needs_fix({"licenses": []})
+
+
+def test_no_licenses_key_needs_fix() -> None:
+    assert needs_fix({})
+
+
+def test_sha256_in_license_id_needs_fix() -> None:
+    assert needs_fix({"licenses": [{"license": {"id": "sha256:abc123"}}]})
+
+
+def test_sha256_in_license_name_needs_fix() -> None:
+    assert needs_fix({"licenses": [{"license": {"name": "sha256:abc123"}}]})
+
+
+def test_sha256_expression_needs_fix() -> None:
+    assert needs_fix({"licenses": [{"expression": "sha256:abc123"}]})
+
+
+def test_valid_spdx_expression_no_fix() -> None:
+    assert not needs_fix({"licenses": [{"expression": "MIT OR Apache-2.0"}]})
+
+
+def test_valid_license_id_no_fix() -> None:
+    assert not needs_fix({"licenses": [{"license": {"id": "MIT"}}]})
+
+
+def test_valid_license_name_no_fix() -> None:
+    assert not needs_fix({"licenses": [{"license": {"name": "MIT"}}]})
 
 
 def test_same_domain_requests_are_spaced() -> None:
